tree: implementation skeleton
This is going to replace the memory management and data parsing in
render.c, which is all over the place and neither very maintainable nor
pretty.

This change is going to allow all kinds of new, better things, for
example a FUSE file system exposing weather information for any program
to read.